requirements.txt
Django==4.1.4
pytest==7.2.0
pytest-django==4.5.2
djangorestframework
dj-config-url
psycopg2-binary
drf-spectacular
pyyaml
celery==5.2.7
redis==4.4.0
flower==1.2.0